Consolidate similar graphql queries
