--- client/lib/api/getSong.graphql
@@ -1,10 +1,0 @@
-query getSong($slug: String!) {
-  song(slug: $slug) {
-    id
-    slug
-    title
-    recommendedKey
-    timeSignature
-    bpm
-  }
-}
