--- client/lib/api/searchSongs.graphql
@@ -1,9 +1,0 @@
-query searchSongs($query: String, $filters: SearchFilters) {
-  searchSongs(query: $query, filters: $filters) {
-    slug
-    title
-    recommendedKey
-    timeSignature
-    bpm
-  }
-}
