--- client/lib/api/updateSong.graphql
@@ -1,5 +1,0 @@
-mutation updateSong($id: ID!, $data: UpdateSongInput!) {
-  updateSong(id: $id, data: $data) {
-    slug
-  }
-}
